feat: implement update review service.
implement update review repository

implement delete and update routes

creating, updating or deleting reviews now updates score of the reviewed movies.

+++ src/jvmMain/kotlin/me/user/application/data/repository/review/ReviewRepositoryImpl.kt
@@ -3,6 +3,7 @@ package me.user.application.data.repository.review
 import io.ktor.http.*
 import me.user.application.data.service.review.ReviewService
 import me.user.application.routes.review.params.CreateReviewParams
+import me.user.application.routes.review.params.UpdateReviewParams
 import me.user.application.utils.BaseResponse
 
 class ReviewRepositoryImpl(
@@ -51,26 +52,51 @@ class ReviewRepositoryImpl(
     override suspend fun createReview(params: CreateReviewParams): BaseResponse<Any> {
         return try {
             val result = reviewService.createReview(params)
-            reviewService.updateMovieScore(result.movieId, result.rating)
+            reviewService.updateMovieScore(result.movieId, result.rating, 1)
             return BaseResponse.SuccessResponse(data = result, statusCode = HttpStatusCode.Created)
         }catch (e: Exception) {
             BaseResponse.ErrorResponse(message = e.message ?: "Error creating review")
         }
     }
 
-    override suspend fun updateReview(params: CreateReviewParams): BaseResponse<Any> {
-        TODO("Not yet implemented")
+    override suspend fun updateReview(params: UpdateReviewParams): BaseResponse<Any> {
+        return try {
+            val review = reviewService.getReview(params.id)
+                ?: return BaseResponse.ErrorResponse(message = "Review not found", statusCode = HttpStatusCode.NotFound)
+
+            if (review.userId != params.userId) {
+                return BaseResponse.ErrorResponse(message = "User not authorized to update review", statusCode = HttpStatusCode.Unauthorized)
+            }
+
+            val result = reviewService.updateReview(review.copy(
+                review = params.review?: review.review,
+                rating = params.score?: review.rating
+            ))
+            if(review.rating != result.rating) {
+                reviewService.updateMovieScore(review.movieId, (result.rating - review.rating), 0)
+            }
+            BaseResponse.SuccessResponse(data = result, statusCode = HttpStatusCode.OK)
+        } catch (e: Exception) {
+            BaseResponse.ErrorResponse(message = e.message ?: "Error updating review", statusCode = HttpStatusCode.InternalServerError)
+        }
     }
 
-    override suspend fun deleteReview(id: Int): BaseResponse<Any> {
+    override suspend fun deleteReview(id: Int, userId: Int): BaseResponse<Any> {
         return try {
+            val review = reviewService.getReview(id)
+                ?: return BaseResponse.ErrorResponse(message = "Review not found", statusCode = HttpStatusCode.NotFound)
+
+            if (review.userId != userId) {
+                return BaseResponse.ErrorResponse(message = "User not authorized to update review", statusCode = HttpStatusCode.Unauthorized)
+            }
             val result = reviewService.deleteReview(id)
             if (result) {
                 BaseResponse.SuccessResponse(statusCode = HttpStatusCode.NoContent)
             } else {
                 BaseResponse.ErrorResponse(message = "Review not found", statusCode = HttpStatusCode.NotFound)
             }
-        } catch (e: Exception) {
+        }
+        catch (e: Exception) {
             BaseResponse.ErrorResponse(message = e.message ?: "Error deleting review", statusCode = HttpStatusCode.InternalServerError)
         }
     }
